# Conflicts:
#	lib/di/injector.config.dart
#	lib/mash/data/local/data_sources/auth_local_data_source.dart
#	lib/mash/domain/entities/auth/auth_response_entity.dart
#	lib/mash/presentation/manager/auth_bloc/auth_bloc.dart

+++ lib/mash/data/remote/models/auth/auth_response_model.dart
@@ -45,7 +45,7 @@ class LoginResTableModel extends LoginResTableEntity {
         studentId: json["STUDENT_ID"],
         admissionNo: json["ADMISSION_NO"],
         studentName: json["STUDENT_NAME"],
-        compId: json["COMP_ID"].toDouble(),
+        compId: json["COMP_ID"],
         roleId: json["ROLE_ID"].toDouble(),
         activeStatus: json["ACTIVE_STATUS"].toDouble(),
         pwdChangeStatus: json["PWD_CHANGE_STATUS"].toDouble(),
